+++ src/control_plane/project_infrastructure/mod.rs
@@ -5,6 +5,7 @@ mod plan_typesense_project_resources;
 mod prepare_project_services;
 mod prepared_project_service;
 mod project_service_preparation_error;
+mod project_service_preparation_strategy;
 
 pub(crate) use plan_meilisearch_project_resources::plan_meilisearch_project_resources;
 pub(crate) use plan_opensearch_project_resources::{
@@ -15,6 +16,7 @@ pub(crate) use plan_typesense_project_resources::plan_typesense_project_resource
 pub(crate) use prepare_project_services::prepare_project_services;
 pub(crate) use prepared_project_service::PreparedProjectService;
 pub(crate) use project_service_preparation_error::ProjectServicePreparationError;
+pub(crate) use project_service_preparation_strategy::ProjectServicePreparationStrategy;
 
 #[cfg(test)]
 mod tests;